Move DataStreamUtils to the datastream API package so that we can actually use it to expose package-private constructors or methods for experimental features.
